- renamed syncManager folder to settings - fixed sms settings design
