+++ app/src/main/java/org/dhis2/data/user/UserComponent.java
@@ -51,8 +51,8 @@ import org.dhis2.usescases.sms.SmsComponent;
 import org.dhis2.usescases.sms.SmsModule;
 import org.dhis2.usescases.sync.SyncComponent;
 import org.dhis2.usescases.sync.SyncModule;
-import org.dhis2.usescases.syncManager.SyncManagerComponent;
-import org.dhis2.usescases.syncManager.SyncManagerModule;
+import org.dhis2.usescases.settings.SyncManagerComponent;
+import org.dhis2.usescases.settings.SyncManagerModule;
 import org.dhis2.usescases.teiDashboard.TeiDashboardComponent;
 import org.dhis2.usescases.teiDashboard.TeiDashboardModule;
 import org.dhis2.usescases.teiDashboard.eventDetail.EventDetailComponent;
