+++ app/src/main/java/org/dhis2/usescases/main/MainActivity.java
@@ -37,8 +37,8 @@ import org.dhis2.usescases.general.FragmentGlobalAbstract;
 import org.dhis2.usescases.jira.JiraFragment;
 import org.dhis2.usescases.main.program.ProgramFragment;
 import org.dhis2.usescases.qrReader.QrReaderFragment;
-import org.dhis2.usescases.syncManager.ErrorDialog;
-import org.dhis2.usescases.syncManager.SyncManagerFragment;
+import org.dhis2.usescases.settings.ErrorDialog;
+import org.dhis2.usescases.settings.SyncManagerFragment;
 import org.dhis2.usescases.teiDashboard.nfc_data.NfcDataWriteActivity;
 import org.dhis2.utils.Constants;
 import org.dhis2.utils.DateUtils;
