+++ app/src/main/java/org/dhis2/usescases/settings/ErrorDialog.java
@@ -1,4 +1,4 @@
-package org.dhis2.usescases.syncManager;
+package org.dhis2.usescases.settings;
 
 import android.app.Dialog;
 import android.content.Context;
