+++ app/src/main/java/org/dhis2/usescases/settings/SyncManagerModule.java
@@ -1,4 +1,4 @@
-package org.dhis2.usescases.syncManager;
+package org.dhis2.usescases.settings;
 
 import org.dhis2.data.dagger.PerFragment;
 import org.hisp.dhis.android.core.D2;
