+++ app/src/main/java/org/dhis2/usescases/settings/SyncManagerPresenter.java
@@ -1,4 +1,4 @@
-package org.dhis2.usescases.syncManager;
+package org.dhis2.usescases.settings;
 
 import android.content.Context;
 import android.content.SharedPreferences;
