+++ app/src/main/res/drawable/gray_border_gray_solid_box.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <stroke android:color="#61333333" android:width="1dp"/>
+    <solid android:color="#80C4C4C4"/>
+
+</shape>
