+++ app/src/main/res/drawable/ic_setting_sms.xml
@@ -1,0 +1,8 @@
+<vector android:height="24dp" android:viewportHeight="40"
+    android:viewportWidth="40" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#00000000"
+        android:pathData="M39.5,20C39.5,9.2304 30.7696,0.5 20,0.5C9.2304,0.5 0.5,9.2304 0.5,20C0.5,30.7696 9.2304,39.5 20,39.5C30.7696,39.5 39.5,30.7696 39.5,20Z"
+        android:strokeColor="#5C5C5C" android:strokeWidth="1"/>
+    <path android:fillColor="#5C5C5C" android:pathData="M10.7715,7.8809C9.5084,7.8809 8.4863,8.9767 8.4863,10.3145L8.4746,32.2031L13.0684,27.3398H20.2363C20.6432,25.7209 21.5786,24.284 22.8942,23.2566C24.2099,22.2293 25.8307,21.6702 27.5,21.668C28.8932,21.6709 30.258,22.0618 31.4414,22.7969V10.3145C31.4414,8.9767 30.4077,7.8809 29.1445,7.8809H10.7715ZM13.2246,10.0215C13.714,10.0215 14.1505,10.1304 14.5332,10.3496C14.9159,10.565 15.2136,10.8714 15.4238,11.2676C15.6372,11.6637 15.7422,12.1134 15.7422,12.6172H14.332C14.332,12.2326 14.2328,11.9342 14.0352,11.7227C13.8375,11.5073 13.5588,11.4004 13.2012,11.4004C12.8561,11.4004 12.5879,11.4892 12.3965,11.6699C12.2051,11.8468 12.1094,12.0827 12.1094,12.375C12.1094,12.6481 12.2206,12.8759 12.4434,13.0605C12.6692,13.2452 13.0014,13.4185 13.4375,13.5801C14.2406,13.8762 14.8244,14.2432 15.1914,14.6816C15.5585,15.1201 15.7422,15.6665 15.7422,16.3203C15.7422,17.0472 15.5189,17.6178 15.0703,18.0332C14.6217,18.4447 14.017,18.6504 13.2578,18.6504C12.7308,18.6504 12.2513,18.5334 11.8184,18.2988C11.3854,18.0604 11.0552,17.7357 10.8262,17.3242C10.6003,16.9127 10.4863,16.4368 10.4863,15.8945H11.9023C11.9023,16.8214 12.3543,17.2832 13.2578,17.2832C13.5935,17.2832 13.8567,17.2005 14.0449,17.0352C14.2332,16.8659 14.3262,16.632 14.3262,16.332C14.3262,16.0051 14.2332,15.7551 14.0449,15.582C13.8567,15.4051 13.5167,15.2196 13.0273,15.0273C12.5379,14.8312 12.1507,14.6396 11.8652,14.4512C11.0872,13.9358 10.6992,13.2422 10.6992,12.3691C10.6992,11.9153 10.8027,11.5101 11.0098,11.1562C11.22,10.7986 11.5192,10.5203 11.9082,10.3203C12.3003,10.1203 12.7384,10.0215 13.2246,10.0215ZM27.6055,10.0215C28.0949,10.0215 28.5313,10.1304 28.9141,10.3496C29.2968,10.565 29.5925,10.8714 29.8027,11.2676C30.0161,11.6637 30.123,12.1134 30.123,12.6172H28.7109C28.7109,12.2326 28.6137,11.9342 28.416,11.7227C28.2184,11.5073 27.9397,11.4004 27.582,11.4004C27.2369,11.4004 26.9687,11.4892 26.7773,11.6699C26.586,11.8468 26.4902,12.0827 26.4902,12.375C26.4902,12.6481 26.6015,12.8759 26.8242,13.0605C27.0501,13.2452 27.3823,13.4185 27.8184,13.5801C28.6215,13.8762 29.2052,14.2432 29.5723,14.6816C29.9393,15.1201 30.123,15.6665 30.123,16.3203C30.123,17.0472 29.8998,17.6178 29.4512,18.0332C29.0026,18.4447 28.3979,18.6504 27.6387,18.6504C27.1116,18.6504 26.6321,18.5334 26.1992,18.2988C25.7663,18.0604 25.436,17.7357 25.207,17.3242C24.9812,16.9127 24.8672,16.4368 24.8672,15.8945H26.2832C26.2832,16.8214 26.7352,17.2832 27.6387,17.2832C27.9743,17.2832 28.2376,17.2005 28.4258,17.0352C28.614,16.8659 28.707,16.632 28.707,16.332C28.707,16.0051 28.614,15.7551 28.4258,15.582C28.2376,15.4051 27.8976,15.2196 27.4082,15.0273C26.9188,14.8312 26.5316,14.6396 26.2461,14.4512C25.4681,13.9358 25.0781,13.2422 25.0781,12.3691C25.0781,11.9153 25.1816,11.5101 25.3887,11.1562C25.5989,10.7986 25.9001,10.5203 26.2891,10.3203C26.6812,10.1203 27.1192,10.0215 27.6055,10.0215ZM16.7031,10.1367H18.5469L20.3066,16.2285L22.0586,10.1367H23.9121V18.5352H22.4961V16.2402L22.6367,12.2773L20.7871,18.5352H19.8184L17.9727,12.2812L18.1152,16.2402V18.5352H16.7031V10.1367ZM10.9746,19.5332H30.5723V20.8477H10.9746V19.5332ZM10.9961,22.1289H23.166C22.5504,22.4934 21.984,22.9351 21.4805,23.4434H10.9961V22.1289ZM10.9961,24.7246H20.4414C20.1665,25.1408 19.9292,25.5807 19.7324,26.0391H10.9961V24.7246Z"/>
+    <path android:fillColor="#5C5C5C" android:pathData="M31.9625,30.085C31.985,29.8975 32,29.7025 32,29.5C32,29.2975 31.985,29.1025 31.955,28.915L33.2225,27.925C33.335,27.835 33.365,27.67 33.2975,27.5425L32.0975,25.465C32.0225,25.33 31.865,25.285 31.73,25.33L30.2375,25.93C29.9225,25.69 29.5925,25.495 29.225,25.345L29,23.755C28.9775,23.605 28.85,23.5 28.7,23.5H26.3C26.15,23.5 26.03,23.605 26.0075,23.755L25.7825,25.345C25.415,25.495 25.0775,25.6975 24.77,25.93L23.2775,25.33C23.1425,25.2775 22.985,25.33 22.91,25.465L21.71,27.5425C21.635,27.6775 21.665,27.835 21.785,27.925L23.0525,28.915C23.0225,29.1025 23,29.305 23,29.5C23,29.695 23.015,29.8975 23.045,30.085L21.7775,31.075C21.665,31.165 21.635,31.33 21.7025,31.4575L22.9025,33.535C22.9775,33.67 23.135,33.715 23.27,33.67L24.7625,33.07C25.0775,33.31 25.4075,33.505 25.775,33.655L26,35.245C26.03,35.395 26.15,35.5 26.3,35.5H28.7C28.85,35.5 28.9775,35.395 28.9925,35.245L29.2175,33.655C29.585,33.505 29.9225,33.3025 30.23,33.07L31.7225,33.67C31.8575,33.7225 32.015,33.67 32.09,33.535L33.29,31.4575C33.365,31.3225 33.335,31.165 33.215,31.075L31.9625,30.085ZM27.5,31.75C26.2625,31.75 25.25,30.7375 25.25,29.5C25.25,28.2625 26.2625,27.25 27.5,27.25C28.7375,27.25 29.75,28.2625 29.75,29.5C29.75,30.7375 28.7375,31.75 27.5,31.75Z"/>
+</vector>
