+++ app/src/main/res/layout/fragment_settings.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="presenter"
-            type="org.dhis2.usescases.syncManager.SyncManagerContracts.Presenter" />
+            type="org.dhis2.usescases.settings.SyncManagerContracts.Presenter" />
 
         <variable
             name="dataSyncMessage"
@@ -26,9 +26,9 @@
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
+            android:background="@color/white"
             android:orientation="vertical"
             android:paddingTop="8dp"
-            android:background="@color/white"
             android:paddingBottom="@dimen/default_top_margin">
 
             <androidx.constraintlayout.widget.ConstraintLayout
@@ -503,20 +503,24 @@
                 android:id="@+id/smsSettings"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:onClick="@{()->presenter.onItemClick(6)}">
+                android:onClick="@{()->presenter.onItemClick(7)}">
 
                 <include
                     android:id="@+id/smsLayout"
                     layout="@layout/fragment_settings_item"
-                    app:icon="@{R.drawable.ic_reset}"
+                    app:icon="@{R.drawable.ic_setting_sms}"
                     app:message="@{@string/settingsSms_descr}"
                     app:title="@{@string/settingsSms}" />
 
                 <FrameLayout
+                    android:id="@+id/smsContent"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="72dp"
-                    app:layout_constraintTop_toBottomOf="@id/smsLayout">
+                    android:layout_marginEnd="16dp"
+                    android:visibility="gone"
+                    app:layout_constraintTop_toBottomOf="@id/smsLayout"
+                    tools:visibility="visible">
 
                     <include
                         android:id="@+id/settings_sms"
